[EFL] API tests failing because Vibration API is using unsigned as duration since r137410 https://bugs.webkit.org/show_bug.cgi?id=104914
Reviewed by Kenneth Rohde Christiansen.

Update the tests to reflect the data type coming from WebCore.

* UIProcess/API/efl/tests/test_ewk2_view.cpp:
(VibrationCbData):
(onVibrate):


Canonical link: https://commits.webkit.org/123140@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@137595 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit2/UIProcess/API/efl/tests/test_ewk2_view.cpp
@@ -802,13 +802,13 @@ struct VibrationCbData {
     bool didReceiveVibrate; // Whether the vibration event received.
     bool didReceiveCancelVibration; // Whether the cancel vibration event received.
     unsigned vibrateCalledCount; // Vibrate callbacks count.
-    uint64_t expectedVibrationTime; // Expected vibration time.
+    unsigned expectedVibrationTime; // Expected vibration time.
 };
 
 static void onVibrate(void* userData, Evas_Object*, void* eventInfo)
 {
     VibrationCbData* data = static_cast<VibrationCbData*>(userData);
-    uint64_t* vibrationTime = static_cast<uint64_t*>(eventInfo);
+    unsigned* vibrationTime = static_cast<unsigned*>(eventInfo);
     if (*vibrationTime == data->expectedVibrationTime)
         data->didReceiveVibrate = true;
     data->vibrateCalledCount++;
